backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
google-genai==1.5.0
mcp==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
